+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added controller interface api's for Policy management.
 - External EDC service interface api updated.
 - Updated supported sub-model implementation classes.
+- EDC asset update refactored in supported submodels.
 
 
 ## [2.3.6] - 2024-03-06